refactor: move editor grid pattern into table feature
