+++ src/feature/table/index.ts
@@ -1,2 +1,3 @@
-export * from './component/Table'
 export * from './store'
+export * from './component/Table'
+export * from './component/RectangularGrid'
